+++ bitcoin/src/p2p/message_blockdata.rs
@@ -13,6 +13,8 @@ use crate::blockdata::block::BlockHash;
 use crate::blockdata::transaction::{Txid, Wtxid};
 use crate::consensus::encode::{self, Decodable, Encodable};
 use crate::internal_macros::impl_consensus_encoding;
+#[cfg(rust_v_1_65)]
+use crate::internal_macros::impl_encoding_from_consensus;
 use crate::p2p;
 
 /// An inventory item.
@@ -96,6 +98,13 @@ impl Decodable for Inventory {
         })
     }
 }
+
+#[cfg(rust_v_1_65)]
+impl_encoding_from_consensus!(Inventory);
+#[cfg(rust_v_1_65)]
+impl_encoding_from_consensus!(GetBlocksMessage);
+#[cfg(rust_v_1_65)]
+impl_encoding_from_consensus!(GetHeadersMessage);
 
 // Some simple messages