+++ consensus_encoding/tests/iter.rs
@@ -49,9 +49,9 @@ impl<const N: usize, const M: usize, const L: usize> Encodable for TestCatArray3
 fn hex_iter() {
     let data = TestArray([255u8, 240, 9, 135]);
     let byte_iter = EncoderByteIter::new(data.encoder());
-    let mut iter = BytesToHexIter::new(byte_iter, hex::Case::Upper);
+    let mut iter = BytesToHexIter::new(byte_iter);
 
-    let expect_str = "FFF00987";
+    let expect_str = "fff00987";
     for byte in expect_str.chars() {
         let iter_byte = iter.next().unwrap();
         assert_eq!(iter_byte, byte);
@@ -64,7 +64,7 @@ fn hex_iter() {
 fn hex_iter_cat_encoder() {
     let data = TestCatArray([222u8, 173], [190u8, 239]);
     let byte_iter = EncoderByteIter::new(data.encoder());
-    let mut iter = BytesToHexIter::new(byte_iter, hex::Case::Lower);
+    let mut iter = BytesToHexIter::new(byte_iter);
 
     let expect_str = "deadbeef";
     for byte in expect_str.chars() {