+++ units/src/lib.rs
@@ -25,6 +25,10 @@ extern crate alloc;
 #[cfg(feature = "std")]
 extern crate std;
 
+/// Bitcoin consensus encoding sans-I/O primitives.
+#[cfg(all(feature = "encoding", rust_v_1_65))]
+pub extern crate encoding;
+
 /// A generic serialization/deserialization framework.
 #[cfg(feature = "serde")]
 pub extern crate serde;